Stable version of the object traking module Needs improvement
