+++ processing/computer_vision/.gitignore
@@ -1,0 +1,1 @@
+/YOLOv5/zips
