+++ processing/computer_vision/YOLOv5/details/bigDataset.json
@@ -1,0 +1,64 @@
+{
+    "model": {
+        "xml": "bigDataset.xml",
+        "bin": "bigDataset.bin"
+    },
+    "nn_config": {
+        "output_format": "detection",
+        "NN_family": "YOLO",
+        "input_size": "416x416",
+        "NN_specific_metadata": {
+            "classes": 5,
+            "coordinates": 4,
+            "anchors": [
+                1.7275390625,
+                2.576171875,
+                2.7109375,
+                3.76953125,
+                3.939453125,
+                5.24609375,
+                5.59375,
+                7.31640625,
+                7.96484375,
+                10.1875,
+                11.28125,
+                14.3046875,
+                16.1875,
+                20.328125,
+                24.09375,
+                30.25,
+                39.28125,
+                50.625
+            ],
+            "anchor_masks": {
+                "side52": [
+                    0,
+                    1,
+                    2
+                ],
+                "side26": [
+                    3,
+                    4,
+                    5
+                ],
+                "side13": [
+                    6,
+                    7,
+                    8
+                ]
+            },
+            "iou_threshold": 0.5,
+            "confidence_threshold": 0.5
+        }
+    },
+    "mappings": {
+        "labels": [
+            "blue_cone",
+            "large_orange_cone",
+            "orange_cone",
+            "unknown_cone",
+            "yellow_cone"
+        ]
+    },
+    "version": 1
+}
